fix: correção ao excluir extensão para não apagar pastas que não estejam vazias

+++ admin/controller/marketplace/install.php
@@ -265,7 +265,7 @@ class ControllerMarketplaceInstall extends Controller {
 
 			if (is_file($file)) {
 				$this->load->model('setting/modification');
-				
+
 				// If xml file just put it straight into the DB
 				$xml = file_get_contents($file);
 
@@ -406,7 +406,7 @@ class ControllerMarketplaceInstall extends Controller {
 			}
 
 			$file = DIR_UPLOAD . $this->session->data['install'] . '.tmp';
-			
+
 			if (is_file($file)) {
 				unlink($file);
 			}
@@ -492,10 +492,10 @@ class ControllerMarketplaceInstall extends Controller {
 					rsort($files);
 
 					foreach ($files as $file) {
-						if (is_file($file)) {
-							unlink($file);
-						} elseif (is_dir($file)) {
-							rmdir($file);
+						if (is_dir($file)) {
+							if ($this->isDirEmpty($file)) {
+								rmdir($file);
+							}
 						}
 					}
 
@@ -504,7 +504,9 @@ class ControllerMarketplaceInstall extends Controller {
 					}
 
 					if (is_dir($source)) {
-						rmdir($source);
+						if ($this->isDirEmpty($source)) {
+							rmdir($source);
+						}
 					}
 				}
 
@@ -524,5 +526,19 @@ class ControllerMarketplaceInstall extends Controller {
 
 		$this->response->addHeader('Content-Type: application/json');
 		$this->response->setOutput(json_encode($json));
+	}
+
+	private function isDirEmpty($dir_name) {
+		if (!is_dir($dir_name)) {
+			return false;
+		}
+
+		foreach (scandir($dir_name) as $dir_file) {
+			if (!in_array($dir_file, array('.','..'))) {
+				return false;
+			}
+		}
+
+		return true;
 	}
 }
